possible fix for issue #27 - permission error 13
in linux, we need to run install with sudo. .goto folder is correctly chowned to the
local user, but running the goto commands is running them as root.
Therefore the one default project we put in the users home folder gets owned by root.

This fix is to add a second chmod after we are done adding magic words.