add end to end tests -- with changes to goto adapted from running them
This adds tests for interacting with the goto commands from the shell.
It allows to set GOTOPATH from environment, in order to run tests
in a different folder then the real goto folder with real projects.

Makefile is updated to add both make unittest, make endtoendtest and make test (both)