fix most issues with zsh
enforced usage of local variables in all shell scripts to fix issue #37.
add more endtoendtests to test initialization.

Concluded with the issue #57 pyenv - that it won't be possible to source
goto through pyenv shims.
The solution will have to be to go away from the strategy of putting all
scripts in path and over to hardcoding absolute paths.. probably