implements feature in #94 - warn about & in url
All write operations that takes uri as input,
now has detection of unescaped ampersand in the url.

That would make goto run in the backround with only half of the url as
argument, and the leftover parts of that url would be attempted executed
by the shell -- most often resulting in "command not found".

This is in fact a security issue, to the unawake user.

http://example.com?a=b&ls

Would execute ls.
ls is harmless. But there are other commands which are not.
Though, this is out of the scope of goto to stop (then we would have to
implement a shell on top of the users shell .. and it is just not what
we are aiming for :P ).

Testing
Turns out the way the commands are executing in the end-to-end framework
is
escaping the ampersand. So it is a bit hard to write automatic tests for
it. Could be done, just a bit unsure how to aproach.

Manual testing seems to work